V7/requirements.txt
streamlit
langchain
langgraph
langchain-groq
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
pymysql
cx-Oracle
pandas
plotly
chromadb
numpy
torch
sentence-transformers
python-dotenv
pydantic
asyncio
scikit-learn
faiss-cpu
tiktoken
boto3
botocore
PyJWT
